renamed manager to factory to reflect its purpose
